extracted internal classes from AetherRepository
